implement line sweeping algorithm

+++ code/02-line/bresenham.cpp
@@ -6,8 +6,8 @@ using namespace SoftwareRenderer;
 void DrawLine(Image& img, Vector2f p1, Vector2f p2, Color color)
 {
     // 将屏幕坐标转换到像素空间
-    auto p1i = Vector2i((int)(p1.x * img.Width()), (int)(p1.y * img.Height()));
-    auto p2i = Vector2i((int)(p2.x * img.Width()), (int)(p2.y * img.Height()));
+    auto p1i = Vector2i(p1.x * img.Width(), p1.y * img.Height());
+    auto p2i = Vector2i(p2.x * img.Width(), p2.y * img.Height());
 
     // 如果斜率大于1，我们将x和y交换，从而统一为横扫描
     auto steep = false;
